Add config mechanism, route service traffic locally
To fix issues with service traffic, we now route that subnet
directly over the relevant ENI

To Support this, we introduce a external configuration mechanism.
This let's is store variable config parameters in a configmap, and
pass it to the relevant places. This also opens up support for
dynamic configuration changes.