Configurable interfaces to route via VPC
When doing outbound comms with masquerade on, all non-VPC local
traffic is masqueraded out the host. This falls down when using
things like VPC peering, as these really need to be routed via the
VPC too. Introduce a configurable set of networks to route via the
VPC gateway, and skip masquerading. Add the RFC1918 networks as
a default set.