Use CNI ptp plugin, rather than custom interface logic
To reduce the amount of nonstandard and difficult to test networking logic
in here, shift to just using the upstream ptp plugin to configure the interfaces.

To do this, cni-vpcnet is adjusted to be a IPAM plugin only. The CNI configuration
generation is switched to output an appropriate config that uses the normal ptp plugin,
with our IPAM. The required routing (container out via ENI, container to container)
rules are moved to be configured statically by address only in the configure tool,
alongside ip masq rules etc.