Handle exhausted IP pools better
Because AWS limits the number of ENIs / IPs to a fairly small number, running
out of addresses is a very real issue. This could be managed with max-pods on
the kubelet, but this won't account for host namespace pods which might make
the small pool even smaller.

As an optional way to handle this, (re-introduce) tainting of full nodes, and
deletion of failed pods.

Tainting is done by regularly checking the IP exhaustion pool, and if full adding
a NoSchedule taint to the the node. This will prevent further pods being added.
Deletion is implemented by hooking in to the CNI allocator. If a pod requests an
IP but the pool is full, this will trigger the daemon do delete the pod. This is
more brutal than I like, but it would likely spin broken for a while. A nice
future enhancement would be seeing if the pod can be "unscheduled"