Refactor controller code
This updates all the controller code to use a shared informer factory.
This gets us closer to "current" best practice, and is also a good starting
point for adding more controller code. The ENI controller is also bought up
to "current" spec. Some basic tests are put in to place as well for
previously untested code.